+++ Lib/Lib/CFileIO.h
@@ -18,8 +18,8 @@ public:
     ~CFileIO();
     int Init();
     void GetSize(const char* szPath, int64_t* Size);
-    int Read(const char* szPath, char* Buffer, long Offset, int64_t Buffer_Size);
-    int Write(const char* szPath, const char* szData, long Offset, int64_t Buffer_Size);
+    int64_t Read(const char* szPath, char* Buffer, long Offset, int64_t Buffer_Size);
+    int64_t Write(const char* szPath, const char* szData, long Offset, int64_t Buffer_Size);
     void Copy(const char* SourceFile, const char* NewFile);
     void SetBuffSize(int64_t Size);
 private:
